refactor: desktop gnb 분리
